faust_app/requirements.txt
faust-streaming
pandas
pymongo
scikit-learn
